requirements.txt
aspy.yaml==1.3.0
cfgv==2.0.1
identify==1.4.9
nodeenv==1.3.4
pre-commit==1.21.0
PyYAML==5.3
six==1.13.0
toml==0.10.0
virtualenv==16.7.9
